% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_table.R
\name{create_table}
\alias{create_table}
\title{it apply
L vector of labour unit
A
K fixed capital}
\usage{
create_table(L, A, K, unit_cost)
}
\description{
it apply
L vector of labour unit
A
K fixed capital
}
