\name{forgnlit30}

\docType{data}

\alias{forgnlit30}

\title{Foreign-born literacy in 1930}

\description{
  This data set contains, on a state level, the proportion of
  residents who are foreign born, and the proportion of residents
  who are literate.  Data come from the 1930 census and were first
  analyzed by Robinson (1950).
}

\usage{data(forgnlit30)}

\format{A data frame containing 3 variables and 48 observations
 \tabular{lll}{
    X \tab numeric \tab proportion of the population that is foreign born \cr
    Y \tab numeric \tab proportion of the population that is illiterate \cr
    FIPS \tab numeric \tab the FIPS state code
  }
}
\references{
 Robinson, W.S. (1950). ``Ecological Correlations and the Behavior
 of Individuals.'' \emph{American Sociological Review}, vol. 15,
 pp.351-357.
}

\keyword{datasets}
