\name{matproii}

\docType{data}

\alias{matproii}

\title{Voter Registration in US Southern States} 

\description{
  This data set contains the voter registration information in four
  southern states in US: Florida, Louisiana, North Carolina, and South
  Carolina.
}

\usage{data(matproii)}

\format{A matrix containing 7 variables and 275 observations}

\references{Gary King (1997) \dQuote{A Solution to the Ecological
    Inference Problem: Reconstructing Individual Behavior from Aggregate
    Data}. Princeton University Press, Princeton, NJ.
}


\keyword{datasets}
