% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.ecoNP.R
\name{print.summary.ecoNP}
\alias{print.summary.ecoNP}
\title{Print the Summary of the Results for the Bayesian Nonparametric Model for Ecological
Inference in 2x2 Tables}
\usage{
\method{print}{summary.ecoNP}(x, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{x}{An object of class \code{summary.ecoNP}.}

\item{digits}{the number of significant digits to use when printing.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
\code{summary.ecoNP} yields an object of class \code{summary.ecoNP}
containing the following elements: 
\item{call}{The call from \code{ecoNP}.}
\item{n.obs}{The number of units.} 
\item{n.draws}{The number of Monte Carlo samples.} 
\item{agg.table}{Aggregate posterior estimates of the marginal
means of \eqn{W_1} and \eqn{W_2} using \eqn{X} and \eqn{N} as weights.} If
\code{param = TRUE}, the following elements are also included:
\item{param.table}{Posterior estimates of model parameters: population mean
estimates of \eqn{W_1} and \eqn{W_2}. If \code{subset} is specified, only a
subset of the population parameters are included.} If \code{unit = TRUE},
the following elements are also included: 
\item{W1.table}{Unit-level posterior estimates for \eqn{W_1}.} 
\item{W2.table}{Unit-level posterior estimates for \eqn{W_2}.}

This object can be printed by \code{print.summary.ecoNP}
}
\description{
\code{summary} method for class \code{ecoNP}.
}
\seealso{
\code{ecoNP}, \code{predict.eco}
}
\keyword{methods}
