% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met.R
\name{read.met.csv2}
\alias{read.met.csv2}
\title{read met file}
\usage{
read.met.csv2(file, skip = NULL, encoding = "UTF-8-BOM")
}
\arguments{
\item{file}{etc}

\item{skip}{etc}

\item{encoding}{etc}
}
\description{
read met file
}
