% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varcov.R
\name{varcov}
\alias{varcov}
\title{Calculate the variance or covariance of the object}
\usage{
varcov(object, ...)
}
\arguments{
\item{object}{An object}

\item{...}{The rest of the input parameters if any}
}
\description{
\code{varcov} returns the variance or covariance of the object.
}
\keyword{methods}
