\name{reg}

\docType{data}

\alias{reg}

\title{Voter Registration in US Southern States} 

\description{
  This data set contains the voter registration information, the voters' racial 
  composition(X) and the registration turnout rate(Y),  in four
  southern states in US: Florida, Louisiana, North Carolina, and South
  Carolina.
}

\usage{data(reg)}

\format{A matrix containing 2 variables and 275 observations}

\references{Gary King (1997) \dQuote{A Solution to the Ecological
    Inference Problem: Reconstructing Individual Behavior from Aggregate
    Data}. Princeton University Press, Princeton, NJ.
}


\keyword{datasets}
