% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg.R
\docType{data}
\name{reg}
\alias{reg}
\title{Voter Registration in US Southern States}
\format{
A data frame containing 5 variables and 275 observations
\tabular{lll}{ X \tab numeric \tab the fraction of Black voters \cr Y \tab
numeric \tab the fraction of voters who registered themselves\cr N \tab
numeric \tab the total number of voters in each county \cr W1 \tab numeric
\tab the actual fraction of Black voters who registered themselves \cr W2
\tab numeric \tab the actual fraction of White voters who registered
themselves }
}
\description{
This data set contains the racial composition, the registration rate, the
number of eligible voters as well as the actual observed racial registration
rates for every county in four US southern states: Florida, Louisiana, North
Carolina, and South Carolina.
}
\references{
King, G. (1997). \dQuote{A Solution to the Ecological Inference
Problem: Reconstructing Individual Behavior from Aggregate Data}. Princeton
University Press, Princeton, NJ.
}
\keyword{datasets}
