\name{reg}

\docType{data}

\alias{reg}

\title{Voter Registration in US Southern States} 

\description{
  This data set contains the racial composition (\eqn{X}), the
  registration rate (\eqn{Y}), and the number of eligible voters
  (\eqn{N}) for every county in four US southern states: Florida,
  Louisiana, North Carolina, and South Carolina.
}

\usage{data(reg)}

\format{A data frame containing 3 variables and 275 observations}

\references{Gary King (1997) \dQuote{A Solution to the Ecological
    Inference Problem: Reconstructing Individual Behavior from Aggregate
    Data}. Princeton University Press, Princeton, NJ.
}

\keyword{datasets}
